Moved the strings tests
