+++ packages/ilib-lint/test/plugins/string/StringFix.test.js
@@ -1,7 +1,7 @@
 /*
  * StringFix.test.js
  *
- * Copyright © 2023, 2025 JEDLSoft
+ * Copyright © 2023 JEDLSoft
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,8 +17,8 @@
  * limitations under the License.
  */
 
-import StringFixCommand from "../../../src/plugins/string/StringFixCommand.js";
-import StringFix from "../../../src/plugins/string/StringFix.js";
+import StringFixCommand from "../src/plugins/string/StringFixCommand.js";
+import StringFix from "../src/plugins/string/StringFix.js";
 
 describe("testStringFix", () => {
     test("StringFixCtorShouldThrowOverlappingCommands", () => {
